Refactor PostPage component to utilize API routes for fetching post and comments, and implement reporting functionality
